+++ src/request_queue.js
@@ -114,13 +114,15 @@ const getRequestId = (uniqueKey) => {
 };
 
 /**
- * A helper class that is used to report results from the
- * [`Apify.utils.puppeteer.enqueueLinks()`](../api/puppeteer#puppeteer.enqueueLinks) function.
+ * A helper class that is used to report results from various
+ * [`RequestQueue`](../api/requestqueue) functions as well as
+ * [`Apify.utils.puppeteer.enqueueLinks()`](../api/puppeteer#puppeteer.enqueueLinks).
  *
- * @typedef {Object} RequestOperationInfo
+ * @typedef {Object} QueueOperationInfo
  * @property {Boolean} wasAlreadyPresent Indicates if request was already present in the queue.
  * @property {Boolean} wasAlreadyHandled Indicates if request was already marked as handled.
  * @property {String} requestId The ID of the added request
+ * @property {Request} request The original `Request` object passed to the `RequestQueue` function.
  */
 
 /**
@@ -155,8 +157,11 @@ const getRequestId = (uniqueKey) => {
  * Each request in the queue is stored as a separate JSON file, where `{STATE}` is either `handled` or `pending`,
  * and `{NUMBER}` is an integer indicating the position of the request in the queue.
  *
- * If the `APIFY_TOKEN` environment variable is provided instead, the data is stored
- * in the <a href="https://www.apify.com/docs/storage#queue" target="_blank">Apify Request Queue</a> cloud storage.
+ * If the `APIFY_TOKEN` environment variable is set but `APIFY_LOCAL_STORAGE_DIR` not, the data is stored in the
+ * <a href="https://www.apify.com/docs/storage#queue" target="_blank">Apify Request Queue</a>
+ * cloud storage. Note that you can force usage of the cloud storage also by passing the `forceCloud`
+ * option to [`Apify.openRequestQueue()`](apify#module_Apify.openRequestQueue) function,
+ * even if the `APIFY_LOCAL_STORAGE_DIR` variable is set.
  *
  * **Example usage:**
  *
@@ -211,12 +216,12 @@ export class RequestQueue {
      *
      * If a request with the same `uniqueKey` property is already present in the queue,
      * it will not be updated. You can find out whether this happened from the resulting
-     * {@link RequestOperationInfo} object.
+     * {@link QueueOperationInfo} object.
      *
-     * @param {Request|Object} request Request object, or an Object to construct a Request from.
+     * @param {Request|Object} request {@link Request} object, or an object to construct a `Request` instance from.
      * @param {Object} [options]
      * @param {Boolean} [options.forefront=false] If `true`, the request will be added to the foremost position in the queue.
-     * @return {RequestOperationInfo}
+     * @return {QueueOperationInfo}
      */
     addRequest(request, options = {}) {
         const { forefront, request: newRequest } = validateAddRequestParams(request, options);
@@ -235,6 +240,8 @@ export class RequestQueue {
                 // request was already handled is there because just one client should be using one queue.
                 wasAlreadyHandled: cachedInfo.isHandled,
                 requestId: cachedInfo.id,
+                // TODO: Why not set request.id to cachedInfo.id???
+                request,
             });
         }
 
@@ -244,16 +251,19 @@ export class RequestQueue {
                 queueId: this.queueId,
                 forefront,
             })
-            .then((requestOperationInfo) => {
-                const { requestId, wasAlreadyHandled } = requestOperationInfo;
+            .then((queueOperationInfo) => {
+                const { requestId, wasAlreadyHandled } = queueOperationInfo;
 
-                this._cacheRequest(cacheKey, requestOperationInfo);
+                this._cacheRequest(cacheKey, queueOperationInfo);
 
                 if (forefront && !this.requestIdsInProgress[requestId] && !wasAlreadyHandled) {
                     this.queueHeadDict.add(requestId, requestId, true);
                 }
 
-                return requestOperationInfo;
+                // TODO: Why not set request.id to cachedInfo.id???
+                queueOperationInfo.request = request;
+
+                return queueOperationInfo;
             });
     }
 
@@ -265,6 +275,8 @@ export class RequestQueue {
      */
     getRequest(requestId) {
         validateGetRequestParams(requestId);
+
+        // TODO: Could we also use requestsCache here? It would be consistent with addRequest()
 
         return requestQueues
             .getRequest({
@@ -309,7 +321,7 @@ export class RequestQueue {
      * Marks request handled after successful processing.
      *
      * @param {Request} request
-     * @return {Promise<RequestOperationInfo>}
+     * @return {Promise<QueueOperationInfo>}
      */
     markRequestHandled(request) {
         validateMarkRequestHandledParams(request);
@@ -325,11 +337,13 @@ export class RequestQueue {
                 request,
                 queueId: this.queueId,
             })
-            .then((requestOperationInfo) => {
+            .then((queueOperationInfo) => {
                 this._removeFromInProgress(request.id);
-                this._cacheRequest(getRequestId(request.uniqueKey), requestOperationInfo);
+                this._cacheRequest(getRequestId(request.uniqueKey), queueOperationInfo);
 
-                return requestOperationInfo;
+                queueOperationInfo.request = request;
+
+                return queueOperationInfo;
             });
     }
 
@@ -342,7 +356,7 @@ export class RequestQueue {
      * @param {Boolean} [options.forefront=false]
      *   If `true` then requests get returned to the start of the queue
      *   and to the back of the queue otherwise.
-     * @return {Promise<RequestOperationInfo>}
+     * @return {Promise<QueueOperationInfo>}
      */
     reclaimRequest(request, options = {}) {
         const { forefront } = validateReclaimRequestParams(request, options);
@@ -353,13 +367,15 @@ export class RequestQueue {
                 queueId: this.queueId,
                 forefront,
             })
-            .then((requestOperationInfo) => {
+            .then((queueOperationInfo) => {
                 this._removeFromInProgress(request.id);
-                this._cacheRequest(getRequestId(request.uniqueKey), requestOperationInfo);
+                this._cacheRequest(getRequestId(request.uniqueKey), queueOperationInfo);
 
                 if (forefront) this.queueHeadDict.add(request.id, request.id, true);
 
-                return requestOperationInfo;
+                queueOperationInfo.request = request;
+
+                return queueOperationInfo;
             });
     }
 
@@ -396,15 +412,15 @@ export class RequestQueue {
      *
      * @ignore
      */
-    _cacheRequest(cacheKey, requestOperationInfo) {
+    _cacheRequest(cacheKey, queueOperationInfo) {
         checkParamOrThrow(cacheKey, 'cacheKey', 'String');
-        checkParamOrThrow(requestOperationInfo, 'requestOperationInfo', 'Object');
-        checkParamOrThrow(requestOperationInfo.requestId, 'requestOperationInfo.requestId', 'String');
-        checkParamOrThrow(requestOperationInfo.wasAlreadyHandled, 'requestOperationInfo.wasAlreadyHandled', 'Boolean');
+        checkParamOrThrow(queueOperationInfo, 'queueOperationInfo', 'Object');
+        checkParamOrThrow(queueOperationInfo.requestId, 'queueOperationInfo.requestId', 'String');
+        checkParamOrThrow(queueOperationInfo.wasAlreadyHandled, 'queueOperationInfo.wasAlreadyHandled', 'Boolean');
 
         this.requestsCache.add(cacheKey, {
-            id: requestOperationInfo.requestId,
-            isHandled: requestOperationInfo.wasAlreadyHandled,
+            id: queueOperationInfo.requestId,
+            isHandled: queueOperationInfo.wasAlreadyHandled,
         });
     }
 
@@ -651,6 +667,7 @@ export class RequestQueueLocal {
                 const queueOrderNo = this._getQueueOrderNo(forefront);
 
                 // Add ID as server does.
+                // TODO: This way of cloning doesn't preserve Dates!
                 const requestCopy = JSON.parse(JSON.stringify(request));
                 requestCopy.id = getRequestId(request.uniqueKey);
 
@@ -662,6 +679,7 @@ export class RequestQueueLocal {
                             requestId: existingRequest.id,
                             wasAlreadyHandled: existingRequest && existingRequest.handledAt,
                             wasAlreadyPresent: true,
+                            request,
                         }));
                 }
 
@@ -675,6 +693,7 @@ export class RequestQueueLocal {
                         requestId: requestCopy.id,
                         wasAlreadyHandled: false,
                         wasAlreadyPresent: false,
+                        request,
                     }));
             });
     }
@@ -750,6 +769,7 @@ export class RequestQueueLocal {
                             requestId: request.id,
                             wasAlreadyHandled: false,
                             wasAlreadyPresent: true,
+                            request,
                         };
                     });
             });
@@ -783,6 +803,7 @@ export class RequestQueueLocal {
                             requestId: request.id,
                             wasAlreadyHandled: false,
                             wasAlreadyPresent: true,
+                            request,
                         };
                     });
             });
@@ -834,15 +855,23 @@ const getOrCreateQueue = (queueIdOrName) => {
  * @param {string} [queueIdOrName]
  *   ID or name of the request queue to be opened. If `null` or `undefined`,
  *   the function returns the default request queue associated with the actor run.
+ * @param {object} [options]
+ * @param {boolean} [options.forceCloud=false]
+ *   If set to `true` then the function uses cloud storage usage even if the `APIFY_LOCAL_STORAGE_DIR`
+ *   environment variable is set. This way it is possible to combine local and cloud storage.
  * @returns {Promise<RequestQueue>}
  * @memberof module:Apify
  * @name openRequestQueue
  */
-export const openRequestQueue = (queueIdOrName) => {
+export const openRequestQueue = (queueIdOrName, options = {}) => {
     checkParamOrThrow(queueIdOrName, 'queueIdOrName', 'Maybe String');
+    checkParamOrThrow(options, 'options', 'Object');
     ensureTokenOrLocalStorageEnvExists('request queue');
 
-    return process.env[ENV_VARS.LOCAL_STORAGE_DIR]
+    const { forceCloud = false } = options;
+    checkParamOrThrow(forceCloud, 'options.forceCloud', 'Boolean');
+
+    return process.env[ENV_VARS.LOCAL_STORAGE_DIR] && !forceCloud
         ? openLocalStorage(queueIdOrName, ENV_VARS.DEFAULT_REQUEST_QUEUE_ID, RequestQueueLocal, queuesCache)
         : openRemoteStorage(queueIdOrName, ENV_VARS.DEFAULT_REQUEST_QUEUE_ID, RequestQueue, queuesCache, getOrCreateQueue);
 };