+++ test/request_list.js
@@ -2,7 +2,7 @@ import 'babel-polyfill';
 import _ from 'underscore';
 import chai, { expect } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import request from 'request-promise';
+import request from 'request-promise-native';
 import sinon from 'sinon';
 import { delayPromise } from 'apify-shared/utilities';
 import { ACTOR_EVENT_NAMES_EX } from '../build/constants';
@@ -513,6 +513,36 @@ describe('Apify.RequestList', () => {
         await requestList.initialize();
 
         expect(requestList.length()).to.be.eql(4);
+    });
+
+    it('it gets correct handledCount()', async () => {
+        const sources = [
+            { url: 'https://www.example.com' },
+            { url: 'https://www.ams360.com' },
+            { url: 'https://www.anybus.com' },
+            { url: 'https://www.anychart.com' },
+            { url: 'https://www.example.com' },
+        ];
+
+        const requestList = new Apify.RequestList({
+            sources,
+        });
+
+        await requestList.initialize();
+
+        const req1 = await requestList.fetchNextRequest();
+        const req2 = await requestList.fetchNextRequest();
+        const req3 = await requestList.fetchNextRequest();
+        expect(requestList.handledCount()).to.be.eql(0);
+
+        await requestList.markRequestHandled(req2);
+        expect(requestList.handledCount()).to.be.eql(1);
+
+        await requestList.markRequestHandled(req3);
+        expect(requestList.handledCount()).to.be.eql(2);
+
+        await requestList.reclaimRequest(req1);
+        expect(requestList.handledCount()).to.be.eql(2);
     });
 
     it('should correctly keep duplicate URLs while keepDuplicateUrls is set', async () => {