+++ website/core/Footer.js
@@ -36,13 +36,13 @@ class Footer extends React.Component {
                     </a>
                     <div>
                         <h5>Docs</h5>
-                        <a href={this.docUrl('guides/motivation.html', this.props.language)}>
+                        <a href={this.docUrl('guides/motivation', this.props.language)}>
                             Guide
                         </a>
-                        <a href={this.docUrl('examples/basiccrawler.html', this.props.language)}>
+                        <a href={this.docUrl('examples/basiccrawler', this.props.language)}>
                             Examples
                         </a>
-                        <a href={this.docUrl('api/apify.html', this.props.language)}>
+                        <a href={this.docUrl('api/apify', this.props.language)}>
                             Reference
                         </a>
                     </div>
@@ -73,7 +73,7 @@ class Footer extends React.Component {
                     </div>
                     <div>
                         <h5>More</h5>
-                        <a href="https://www.apify.com" target="_blank">Apify Cloud Platform</a>
+                        <a href="https://www.apify.com" target="_blank">Apify Cloud</a>
                         <a href="https://docusaurus.io" target="_blank">Docusaurus</a>
                         <a href={this.props.config.repoUrl} target="_blank">GitHub</a>
                     </div>